[16] (feat on refactor) BIG REFACTOR
