+++ src/cleaner.rs
@@ -20,7 +20,6 @@ use crate::project::{Project, Projects};
 /// The `Cleaner` struct provides methods for removing build directories
 /// (such as `target/` for Rust projects and `node_modules/` for Node.js projects)
 /// with parallel processing, progress reporting, and comprehensive error handling.
-#[allow(dead_code)]
 pub struct Cleaner;
 
 impl Cleaner {
@@ -37,7 +36,6 @@ impl Cleaner {
     /// let cleaner = Cleaner::new();
     /// ```
     #[must_use]
-    #[allow(dead_code)]
     pub fn new() -> Self {
         Self
     }
@@ -89,7 +87,6 @@ impl Cleaner {
     /// Individual project cleanup failures do not stop the overall process.
     /// All errors are collected and reported at the end, allowing the
     /// cleanup to proceed for projects that can be successfully processed.
-    #[allow(dead_code)]
     pub fn clean_projects(projects: Projects) {
         let total_projects = projects.len();
         let total_size: u64 = projects.get_total_size();
@@ -223,7 +220,6 @@ impl Cleaner {
 ///     Err(e) => eprintln!("Cleanup failed: {}", e),
 /// }
 /// ```
-#[allow(dead_code)]
 fn clean_single_project(project: &Project) -> Result<u64> {
     let build_dir = &project.build_arts.path;
 
@@ -276,7 +272,6 @@ fn clean_single_project(project: &Project) -> Result<u64> {
 /// let size = calculate_directory_size(Path::new("/path/to/directory"));
 /// println!("Directory size: {} bytes", size);
 /// ```
-#[allow(dead_code)]
 fn calculate_directory_size(path: &std::path::Path) -> u64 {
     let mut total_size = 0u64;
