+++ src/utils/mod.rs
@@ -1,0 +1,8 @@
+//! Utility functions and helpers.
+//!
+//! This module contains utility functions used throughout the application,
+//! such as size parsing and formatting helpers.
+
+pub mod size;
+
+pub use size::parse_size;
